requirements.txt
python-telegram-bot==21.3
httpx==0.27.0
pydantic==2.8.2
PyYAML==6.0.2
pytrends==4.9.2
pandas==2.2.2
numpy==1.26.4
aiohttp==3.9.5
